Added extra validation for orders form.

+++ CarDealership/Forms/Orders.cs
@@ -22,17 +22,39 @@ namespace CarDealership.Forms
             ordersGridView.DataSource = _orderRepository.Entities.ToTableData();
         }
 
-        private void toggleOrderApproving_Click(object sender, EventArgs e)
+        private void approveButton_Click(object sender, EventArgs e)
         {
-            var orderToToggleApproving = SelectedOrder;
-            orderToToggleApproving.IsApproved = !orderToToggleApproving.IsApproved;
-            _orderRepository.Update(orderToToggleApproving);
-            ordersGridView.DataSource = _orderRepository.Entities.ToTableData();
+            ToggleOrderApproving(true, "Оберіть заявку для підтвердження");
+        }
+
+        private void rejectButton_Click(object sender, EventArgs e)
+        {
+            ToggleOrderApproving(true, "Оберіть заявку для відхилення");
         }
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
+            if (!ordersGridView.IsSelectedRow())
+            {
+                MessageUtil.ShowError("Оберіть заявку для видалення");
+                return;
+            }
+
             _orderRepository.Delete(SelectedOrder);
+            ordersGridView.DataSource = _orderRepository.Entities.ToTableData();
+        }
+
+        private void ToggleOrderApproving(bool isApproved, string message)
+        {
+            if (!ordersGridView.IsSelectedRow())
+            {
+                MessageUtil.ShowError(message);
+                return;
+            }
+
+            var orderToToggleApproving = SelectedOrder;
+            orderToToggleApproving.IsApproved = isApproved;
+            _orderRepository.Update(orderToToggleApproving);
             ordersGridView.DataSource = _orderRepository.Entities.ToTableData();
         }
     }
